Move preferences to qiskit.aqua
