+++ qiskit_aqua_ui/browser/command_line.py
@@ -27,8 +27,7 @@ def set_preferences_logging():
     Update logging setting with latest external packages
     """
     from qiskit.aqua._logging import get_logging_level, build_logging_config, set_logging_config
-    from qiskit_aqua_cmd.preferences import Preferences
-    preferences = Preferences()
+    preferences = UIPreferences()
     logging_level = logging.INFO
     if preferences.get_logging_config() is not None:
         set_logging_config(preferences.get_logging_config())
