+++ qiskit_aqua_ui/run/_aquaguiprovider.py
@@ -55,11 +55,6 @@ class AquaGUIProvider(GUIProvider):
 
         return self._controller
 
-    def create_preferences(self):
-        """Creates provider preferences."""
-        from qiskit_aqua_cmd import Preferences
-        return Preferences()
-
     def create_uipreferences(self):
         """Creates provider UI preferences."""
         return UIPreferences()
