+++ qiskit_aqua_ui/run/_credentialsview.py
@@ -32,7 +32,7 @@ class CredentialsView(ttk.Frame):
 
         self.pack(fill=tk.BOTH, expand=tk.TRUE)
 
-        from qiskit_aqua_cmd import Preferences
+        from qiskit.aqua import Preferences
         preferences = Preferences()
         self._credentials_preferences = preferences.credentials_preferences
 
@@ -101,7 +101,7 @@ class CredentialsView(ttk.Frame):
         urls = [credentials.url for credentials in self._credentials_preferences.get_all_credentials()]
         dialog = URLEntryDialog(self.master, self)
         dialog.do_init(tk.LEFT)
-        from qiskit_aqua_cmd._credentialspreferences import CredentialsPreferences
+        from qiskit.aqua._credentialspreferences import CredentialsPreferences
         if CredentialsPreferences.URL not in urls:
             dialog._url.insert(0, CredentialsPreferences.URL)
         dialog.do_modal()
@@ -424,8 +424,7 @@ class ProxiesPage(ToolbarView):
         return True
 
     def apply(self, preferences):
-        preferences.set_proxy_urls(
-            self._proxy_urls if len(self._proxy_urls) > 0 else None)
+        preferences.set_proxy_urls(self._proxy_urls if len(self._proxy_urls) > 0 else None)
 
 
 class URLPopup(EntryCustom):
